Fix tock, wire up HUD, cleanup
